refactor(exec): rename `exec*` files into `run*`
