+++ packages/exec/lib/command.js
@@ -1,7 +1,7 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.Command = void 0;
-const exec_1 = require("./exec");
+const run_1 = require("./run");
 /** A helper for running a command */
 class Command {
     /**
@@ -19,7 +19,7 @@ class Command {
      * @returns a command run result
      */
     async run(...args) {
-        return (0, exec_1.run)(this.command, ...this.args.concat(args));
+        return (0, run_1.run)(this.command, ...this.args.concat(args));
     }
     /**
      * Runs the command silently
@@ -27,7 +27,7 @@ class Command {
      * @returns a command run result
      */
     async runSilently(...args) {
-        return (0, exec_1.runSilently)(this.command, ...this.args.concat(args));
+        return (0, run_1.runSilently)(this.command, ...this.args.concat(args));
     }
     /**
      * Runs the command and gets the output
@@ -35,7 +35,7 @@ class Command {
      * @returns a command run result
      */
     async output(...args) {
-        return (0, exec_1.output)(this.command, ...this.args.concat(args));
+        return (0, run_1.output)(this.command, ...this.args.concat(args));
     }
     /**
      * Runs the command silently and gets the output
@@ -43,7 +43,7 @@ class Command {
      * @returns a command run result
      */
     async outputSilently(...args) {
-        return (0, exec_1.outputSilently)(this.command, ...this.args.concat(args));
+        return (0, run_1.outputSilently)(this.command, ...this.args.concat(args));
     }
 }
 exports.Command = Command;
