+++ packages/exec/lib/index.d.ts
@@ -1,3 +1,3 @@
 export { Command } from "./command";
-export { output, outputSilently, run, runSilently } from "./exec";
 export { Result } from "./result";
+export { output, outputSilently, run, runSilently } from "./run";
