+++ packages/exec/lib/run.js
@@ -86,4 +86,4 @@ async function outputSilently(command, ...args) {
     return outputHelper(true, command, ...args);
 }
 exports.outputSilently = outputSilently;
-//# sourceMappingURL=exec.js.map
+//# sourceMappingURL=run.js.map
